requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
drf-yasg==1.21.10
gunicorn==23.0.0
inflection==0.5.1
packaging==24.2
pillow==11.1.0
psycopg==3.2.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
whitenoise==6.9.0
